+++ dhara/dhara/dhara/error.c
@@ -28,7 +28,8 @@ const char *dhara_strerror(dhara_error_t err)
         [DHARA_E_JOURNAL_FULL] = "Journal is full",
         [DHARA_E_NOT_FOUND] = "No such sector",
         [DHARA_E_MAP_FULL] = "Sector map is full",
-        [DHARA_E_CORRUPT_MAP] = "Sector map is corrupted"
+	[DHARA_E_CORRUPT_MAP] = "Sector map is corrupted",
+	[DHARA_E_PAGE_RELIEF] = "Physical page skipped for wear relief"
     };
     const char *msg = NULL;